refactoring resolvers user not defined error occurs
